orientation~>commonline, breakout filter_to_fb_mat, io~>storage
adds missing covariance __init__
