+++ src/aspire/abinitio/commonline_sdp.py
@@ -1,16 +1,14 @@
 import logging
 
-from aspire.orientation import CLOrient3D
+from aspire.abinitio import CLOrient3D
 
 logger = logging.getLogger(__name__)
 
 
-class CommLineLUD(CLOrient3D):
+class CommLineSDP(CLOrient3D):
     """
-    Define a derived class to estimate 3D orientations using Least Unsquared Deviations described as below:
-    L. Wang, A. Singer, and  Z. Wen, Orientation Determination of Cryo-EM Images Using Least Unsquared Deviations,
-    SIAM J. Imaging Sciences, 6, 2450-2483 (2013).
-
+    Class to estimate 3D orientations using Semi-Definite Programming
+    :cite:`DBLP:journals/siamis/SingerS11`
     """
 
     def __init__(self, src):
