+++ src/aspire/covariance/__init__.py
@@ -1,0 +1,2 @@
+from .covar import CovarianceEstimator
+from .covar2d import BatchedRotCov2D, RotCov2D
