+++ src/aspire/source/__init__.py
@@ -4,8 +4,11 @@ import numpy as np
 import pandas as pd
 from scipy.spatial.transform import Rotation as R
 
-from aspire.image import Image, normalize_bg
-from aspire.source.xform import (
+from aspire.image import (
+    Image,    
+    normalize_bg,
+)   
+from aspire.image.xform import (
     Downsample,
     FilterXform,
     IndexedXform,
